Begin ValueView meta-key special processing
When dragging a ValueView with the meta-key pressed, it will
increment (or decrement) its value based on the drag offset.

The communication mechanism for this is set up in this commit.
Still to do is the actual mutation of the ValueView's Value. For
that, we'll need to add a few methods to Value for this to work.